pyproject.toml
[project]
name = "nt-etl-order-book"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cryptography>=46.0.3",
    "polars>=1.35.1",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "rich>=14.2.0",
]
